chore: move global context package out of engine
