+++ pkg/controllers/globalcontext/controller.go
@@ -11,9 +11,9 @@ import (
 	"github.com/kyverno/kyverno/pkg/clients/dclient"
 	"github.com/kyverno/kyverno/pkg/controllers"
 	"github.com/kyverno/kyverno/pkg/engine/adapters"
-	"github.com/kyverno/kyverno/pkg/engine/globalcontext/externalapi"
-	"github.com/kyverno/kyverno/pkg/engine/globalcontext/k8sresource"
-	"github.com/kyverno/kyverno/pkg/engine/globalcontext/store"
+	"github.com/kyverno/kyverno/pkg/globalcontext/externalapi"
+	"github.com/kyverno/kyverno/pkg/globalcontext/k8sresource"
+	"github.com/kyverno/kyverno/pkg/globalcontext/store"
 	controllerutils "github.com/kyverno/kyverno/pkg/utils/controller"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime/schema"
