+++ pkg/globalcontext/invalid/entry.go
