+++ pkg/globalcontext/k8sresource/entry.go
