+++ pkg/globalcontext/store/entry.go
